+++ src/main/java/com/vonage/client/application/DeleteApplicationMethod.java
@@ -56,4 +56,9 @@ class DeleteApplicationMethod extends AbstractMethod<String, Void> {
 
         return null;
     }
+
+    @Override
+    protected RequestBuilder applyAuth(RequestBuilder request) throws VonageClientException {
+        return getAuthMethod(getAcceptableAuthMethods()).applyAsBasicAuth(request);
+    }
 }