+++ src/main/java/com/vonage/client/application/GetApplicationMethod.java
@@ -57,4 +57,9 @@ class GetApplicationMethod extends AbstractMethod<String, Application> {
 
         return Application.fromJson(new BasicResponseHandler().handleResponse(response));
     }
+
+    @Override
+    protected RequestBuilder applyAuth(RequestBuilder request) throws VonageClientException {
+        return getAuthMethod(getAcceptableAuthMethods()).applyAsBasicAuth(request);
+    }
 }