+++ src/main/java/com/vonage/client/application/ListApplicationsMethod.java
@@ -69,4 +69,9 @@ class ListApplicationsMethod extends AbstractMethod<ListApplicationRequest, Appl
 
         return ApplicationList.fromJson(new BasicResponseHandler().handleResponse(response));
     }
+
+    @Override
+    protected RequestBuilder applyAuth(RequestBuilder request) throws VonageClientException {
+        return getAuthMethod(getAcceptableAuthMethods()).applyAsBasicAuth(request);
+    }
 }