+++ src/test/java/com/vonage/client/application/CreateApplicationMethodTest.java
@@ -17,13 +17,14 @@ package com.vonage.client.application;
 
 import com.vonage.client.HttpConfig;
 import com.vonage.client.HttpWrapper;
+import com.vonage.client.auth.TokenAuthMethod;
 import org.apache.http.client.methods.RequestBuilder;
 import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class CreateApplicationMethodTest {
+public class CreateApplicationMethodTest extends AppBasicAuthTest {
     private CreateApplicationMethod method;
 
     @Before