+++ src/test/java/com/vonage/client/application/GetApplicationMethodTest.java
@@ -23,7 +23,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class GetApplicationMethodTest {
+public class GetApplicationMethodTest extends AppBasicAuthTest {
     private GetApplicationMethod method;
 
     @Before